Use a custom enum to represent scopes for clarity
Using `Option<_>` made reading the code give wrong impressions on what
`None` means when talking about scopes. This has been clarified by
using a custom enum `Scopes` with `Scopes::All` replacing the previous
use of `Option::None` to indicate disabling authorization checks.
