+++ src/auth.rs
@@ -1,0 +1,40 @@
+//! Authentication and authorization data structures
+
+use std::collections::BTreeSet;
+use iron;
+use hyper;
+
+/// Authorization scopes.
+#[derive(Clone, Debug, PartialEq)]
+pub enum Scopes {
+    /// Some set of scopes.
+    Some(BTreeSet<String>),
+    /// All possible scopes, authorization checking disabled.
+    All,
+}
+
+/// Storage of authorization parameters for an incoming request, used for
+/// REST API authorization.
+#[derive(Clone, Debug, PartialEq)]
+pub struct Authorization {
+    pub subject: String,
+    pub scopes: Scopes,
+}
+impl iron::typemap::Key for Authorization {
+    type Value = Authorization;
+}
+
+/// Storage of raw authentication data, used both for storing incoming
+/// request authentication, and for authenticating outgoing client requests.
+#[derive(Clone, Debug, PartialEq)]
+pub enum AuthData {
+    /// HTTP Basic auth.
+    Basic(hyper::header::Basic),
+    /// HTTP Bearer auth, used for OAuth2.
+    Bearer(hyper::header::Bearer),
+    /// Header-based or query parameter-based API key auth.
+    ApiKey(String),
+}
+impl iron::typemap::Key for AuthData {
+    type Value = AuthData;
+}
